+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ** Updated: June 24nd, 2025 **
 
-- Added `vscode.nix` with configured plugins for nix, bash, json,toml and VIM
-  keybinds
+- Added `vscode.nix` with configured plugins
+- Thanks to `delciak` for the NIX code for `vscode.nix`
 - Fixed 'hm-find' script not being built
 - Fixed `css` formatting thanks to `mister_simon` for the fix
 - Set `TERMINAL` and `XDG_TERMINAL_EMULATOR` to kitty in `env.nix`